+++ scripts/calculateKpi/protocols/tetherV0/parseReferralTag/alignMagicPrefixAndPadHex.ts
