+++ scripts/calculateKpi/protocols/tetherV0/parseReferralTag/parseReferral.ts
